+++ src/ampmesh/triangle/TriangleMeshIterator.h
@@ -10,7 +10,7 @@ namespace AMP {
 namespace Mesh {
 
 
-template<size_t NG, size_t NP>
+template<uint8_t NG, uint8_t NP>
 class TriangleMesh;
